# openapi/spec.yaml
openapi: 3.1.0
info:
  title: Node API Gateway
  version: 1.0.0
  description: Hono-based API gateway aggregating WooCommerce and Algolia

servers:
  - url: http://localhost:3001
    description: Local development

security:
  - bearerAuth: []
  - apiKeyAuth: []

paths:
  /health:
    get:
      summary: Liveness probe
      security: []
      responses:
        '200':
          description: Service is alive
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: ok }
                  timestamp: { type: string, format: date-time }

  /ready:
    get:
      summary: Readiness probe
      security: []
      responses:
        '200':
          description: Service is ready
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
                  services: { type: object }

  /products:
    get:
      summary: List products
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: search
          in: query
          schema: { type: string }
        - name: category
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Product list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Product' }
                  pagination: { $ref: '#/components/schemas/Pagination' }

  /products/{id}:
    get:
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Product
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }

  /search:
    get:
      summary: Search products via Algolia
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer }
        - name: hitsPerPage
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Search results

  /cart/{cartId}:
    get:
      summary: Get cart
      parameters:
        - name: cartId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Cart

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key

  schemas:
    Product:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        price: { type: number }
        description: { type: string }

    Pagination:
      type: object
      properties:
        page: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }

    ProblemDetails:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }